chore(ci): tag code before creating release
It's a bit clunky, but at the moment this seems to be the best way to
push a tag with Jenkins. There is a set of credentials that supplies the
username and password for a user that has permissions to commit to this
repo.

For more info, see this post on Stackoverflow:
https://stackoverflow.com/questions/33570075/tag-a-repo-from-a-jenkins-workflow-script#

Using the `withCredentials` block guarantees that Jenkins won't output
the variables to the log.