chore(ci): build in container for linux
Provides a container definition for building safe-cli. Unfortunately
when you then subsequently run another build in the container, it is
performing re-compilation. I've only observed this behaviour on this
project, and don't really know what the reason might be.

There was also a little change here in the testing code, which was using
a reference to safe_cli that was assuming it was in a target directory
that was relative to the root of the repository. For a build in a
container this is not the case, so the code has been changed to check if
the CARGO_TARGET_DIR variable is set, and if so use that, otherwise it
will default to a directory at the root of the repo.