style(ci): use java style
Since my main language is Python, I have a habit of using the styles
from that language. I've changed the code to conform with the Java
camelCase styles.

Maybe it's not so much of a big deal in this context, but since the
Groovy standard library functions use the Java style, I'd rather have
the code be consistent with that.