chore(ci): use musl on linux
Updates the container to use musl to build. This required changing the
container from Stretch to the new Buster release. On Stretch I was
seeing strange errors relating to PIC:
```
/usr/bin/ld: /target/release/deps/librust_sodium_sys-1fda9c1022584d80.rlib(libsodium_la-randombytes.o): relocation R_X86_64_32S against symbol `randombytes_sysrandom_implementation' can not be used when making a shared object; recompile with -fPIC
```

The stuff about setting up openssl is necessary because rust-openssl is
a dependency, and when you changed to using musl it wouldn't build
without these things setup.