chore(ci): clean release builds
We found a problem with the last Linux build that was released - it had
the wrong version number. This seems to have happened because Cargo
didn't invalidate the cache that was in the container. When it comes to
builds that are released to the public, we can't really take any chances
with this - we need to prioritise correctness over speed.

This adds a new `build-clean` target that will be invoked whenever we're
on the master branch. There's also another little improvement to run
`strip` against the release builds, to reduce the size a bit.