chore(ci): test runs for safe-api
Introduces running the tests for safe-api in the first stage of the
pipeline.

A new container has been introduced for the api build, and targets that
were previously named generically, e.g. `build`, now have the component
name, e.g. `build-cli`, since this repository is now a Rust workspace.
The containers are now using the component name as a tag.

The `all-tests` script is refactored here to extract each test suite run
into its own script. This is so we can run the cli and api tests
separately in parallel.

There's also a little change here related to the Powershell cleanup
script. For no apparent reason I started getting this:
The argument 'resourcestest-scriptscleanup.ps1' to the -File parameter does not exist.

So there's some kind of parsing issue with the backslashes. I've changed
it back to the Unix-style path separator and this seems to settle it.
Really not sure how this came about. Perhaps it might be because the new
slaves have a newer version of Git Bash that doesn't allow the
Windows-style paths.