refactor(ci): use a generic build target
With so many targets to build, and potentially with 'clean' and 'dev'
targets for each, there was beginning to be a huge proliferation of
build targets that were tedious to define and maintain.

I created one target named 'build-component' that expects a few
environment variables to be defined, then passes the values of those
variables to a little script (the Makefile is just too messy for that
kind of code).

All the `build-X` targets have now been removed in favour of this.

The containers now need to follow this convention:
maidsafe/safe-build-cli:<component>-<target>

For example:
maidsafe/safe-build-cli:cli-x86_64-unknown-linux-gnu

This is much easier to use and maintain. Likewise I'm going to do the
same for the `build-container` targets, which are also proliferating.