chore(ci): update CI real network test to use "latest" node
The node version was explicitly stated in CI, meaning it would need
to be updated every time there was a new release. Easier to point
to "latest" which is stored on S3.
This PR also removes a now redundant check against latest master
node branch - with CD now in sn_node any merges to master are released
as "latest".